breaks out the schema & various functions for surveys into separate files
