+++ src/routes/demographics/Content.svelte
@@ -3,225 +3,14 @@
    * License, v. 2.0. If a copy of the MPL was not distributed with this
    * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-  import { writable } from "svelte/store";
   import { fly } from "svelte/transition";
   import ClearAnswerButton from './ClearAnswerButton.svelte';
-
-  function zipcodeIsValid() {
-    return function zipcodeValid(value) {
-      if (Number.isNaN(Number(value)) || value.length > 5)
-        return "Please enter a five-digit zip code.";
-      else return true;
-    };
-  }
-
-  function buildValidator(validators) {
-    return function validate(value, dirty) {
-      if (!validators || validators.length === 0) {
-        return { dirty, valid: true };
-      }
-
-      const failing = validators.find((v) => v(value) !== true);
-      return {
-        dirty,
-        valid: !failing,
-        message: failing && failing(value),
-      };
-    };
-  }
-
-  export function createFieldValidator(...validators) {
-    const { subscribe, set } = writable({
-      dirty: false,
-      valid: false,
-      message: null,
-    });
-    const validator = buildValidator(validators);
-
-    function action(node, binding) {
-      function validate(value, dirty) {
-        const result = validator(value, dirty);
-        set(result);
-      }
-
-      validate(binding, false);
-
-      return {
-        update(value) {
-          validate(value, true);
-        },
-      };
-    }
-
-    return [{ subscribe }, action];
-  }
+  import schema from './survey-schema';
+  import { questionIsAnswered, clearAnswer, createFieldValidator, zipcodeIsValid, createResultObject } from './survey-tools';
 
   const [zipValidity, validateZip] = createFieldValidator(zipcodeIsValid());
 
-  // add survey schema here.
-  const schema = {
-    age: {
-      key: "age",
-      label: "1. What is your age?",
-      type: "single",
-      columns: true,
-      values: [
-        { key: "19_24", label: "19-24 years old" },
-        { key: "25_34", label: "25-34 years old" },
-        { key: "35_44", label: "35-44 years old" },
-        { key: "45_54", label: "45-54 years old" },
-        { key: "55_64", label: "55-64 years old" },
-        { key: "over_65", label: "65 years and older" },
-      ],
-    },
-    gender: {
-      key: "gender",
-      label: "2. What is your gender?",
-      type: "single",
-      columns: true,
-      values: [
-        { key: "male", label: "Male" },
-        { key: "female", label: "Female" },
-        { key: "neither", label: "Neither choice describes me" },
-        { key: "decline", label: "Decline to identify" },
-      ],
-    },
-    hispanicLatinxSpanishOrigin: {
-      key: "hispanicLatinxSpanishOrigin",
-      label: "3. Are you of Hispanic, Latinx, or Spanish origin?",
-      type: "single",
-      values: [
-        { key: "other", label: "No, not of Hispanic, Latinx, or Spanish origin" },
-        { key: "hispanicLatinxSpanish", label: "Yes" },
-      ],
-    },
-    race: {
-      key: "race",
-      label: "4. What is your race?",
-      type: "multi",
-      columns: true,
-      values: [
-        {
-          key: "american_indian_or_alaska_native",
-          label: "American Indian or Alaska Native",
-        },
-        { key: "asian_indian", label: "Asian Indian" },
-        {
-          key: "black_or_african_american",
-          label: "Black or African American",
-        },
-        { key: "chamorro", label: "Chamorro" },
-        { key: "chinese", label: "Chinese" },
-        { key: "filipino", label: "Filipino" },
-        { key: "japanese", label: "Japanese" },
-        { key: "korean", label: "Korean" },
-        { key: "native_hawaiian", label: "Native Hawaiian" },
-        { key: "samoan", label: "Samoan" },
-        { key: "vietnamese", label: "Vietnamese" },
-        { key: "white", label: "White" },
-        { key: "other_asian", label: "Other Asian" },
-        { key: "other_pacific_islander", label: "Other Pacific Islander" },
-        { key: "some_other_race", label: "Some other race" },
-      ],
-    },
-
-    school: {
-      key: "school",
-      label: "5. What is the highest level of school you have completed?",
-      type: "single",
-      values: [
-        { key: "less_than_high_school", label: "Less than high school" },
-        { key: "some_high_school", label: "Some high school" },
-        {
-          key: "high_school_graduate_or_equivalent",
-          label: "High school grduate or equivalent (for example GED)",
-        },
-        {
-          key: "some_college_but_no_degree_or_in_progress",
-          label: "Some college, but degree not received or in progress",
-        },
-        {
-          key: "associates_degree",
-          label: "Associate's degree (for example AA, AS)",
-        },
-        {
-          key: "bachelors_degree",
-          label: "Bachelor's degree (for example BA, BS, BB)",
-        },
-        {
-          key: "graduate_degree",
-          label:
-            "Graduate degree (for example master's, professional, doctorate)",
-        },
-      ],
-    },
-    income: {
-      key: "income",
-      label:
-        "6. What is your household's combined annual income during the past 12 months?",
-      type: "single",
-      columns: 3,
-      values: [
-        { key: "0_24999", label: "$0 - $24,999" },
-        { key: "25000_49999", label: "$25,000 - $49,999" },
-        { key: "50000_74999", label: "50,000 - $74,999" },
-        { key: "75000_99999", label: "$75,000 - $99,999" },
-        { key: "100000_149999", label: "$100,000 - $149,999" },
-        { key: "ge_150000", label: "$150,000 or more" },
-      ],
-    },
-    zipcode: {
-      key: "zipcode",
-      label: "7. What is your zip code?",
-      type: "text",
-    },
-  };
-
-  function questionIsAnswered(answer, questionType) {
-      if (questionType === 'text') return answer.length > 0;
-      if (questionType === 'single') return answer !== undefined;
-      if (questionType === 'multi') return answer.length > 0;
-  }
-
-  function clearAnswer(question) {
-    switch (schema[question].type) {
-      case "text": {
-        results[question] = '';
-        break;
-      } case "single": {
-        results[question] = undefined;
-        break;
-      } case "multi": {
-        results[question] = [];
-        break;
-      } default: {
-        throw Error('question must have a type');
-      }
-    }
-  }
-
-  export let results = Object.values(schema).reduce((acc, config) => {
-    let defaultValue = undefined;
-    switch (config.type) {
-      case "single": {
-        defaultValue = undefined;
-        break;
-      }
-      case "multi": {
-        defaultValue = [];
-        break;
-      }
-      case "text": {
-        defaultValue = "";
-        break;
-      }
-      default:
-        break;
-    }
-
-    acc[config.key] = defaultValue;
-    return acc;
-  }, {});
+  export let results = createResultObject(schema);
 </script>
 
 <style>
@@ -334,7 +123,7 @@
           {#if questionIsAnswered(results[question], schema[question].type)}
             <ClearAnswerButton on:click={(e) => {
               e.preventDefault();
-              clearAnswer(question);
+              results[question] = clearAnswer(schema[question].type);
             }} />
           {/if}
         </legend>
